Use Array.from on the factory.
We need to be able to re-iterate across state when we get a new state,
so save the state to an Array via Array.from(). The downside is that
we have to wait for the factory to finish producing values before we
can return anything, which could be important for some use cases.

Possible enhancement: create a "tee"'ing generator that yields, while
also storing entries into an array, which would allow us to have real
time output.